renamed top level package from io to com so that it matches group id (to be used with sonatype maven repo)
