+++ src/main/java/com/github/lukaszbudnik/guice/multibindinginterceptor/ioc/Secondary.java
@@ -1,4 +1,4 @@
-package io.github.lukaszbudnik.guice.multibindinginterceptor.ioc;
+package com.github.lukaszbudnik.guice.multibindinginterceptor.ioc;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
