+++ src/test/java/com/github/lukaszbudnik/guice/multibindinginterceptor/ioc/test/StorageService2Impl.java
@@ -1,30 +1,30 @@
-package io.github.lukaszbudnik.guice.multibindinginterceptor.ioc.test;
+package com.github.lukaszbudnik.guice.multibindinginterceptor.ioc.test;
 
+import com.github.lukaszbudnik.guice.multibindinginterceptor.ioc.Primary;
 import com.google.inject.Singleton;
-import io.github.lukaszbudnik.guice.multibindinginterceptor.ioc.Secondary;
 import lombok.extern.log4j.Log4j2;
 
-@Secondary
+@Primary
 @Singleton
 @Log4j2
-public class StorageService1Impl extends AbstractTestStorageService implements StorageService {
+public class StorageService2Impl extends AbstractTestStorageService implements StorageService {
 
     @Override
     public int put(String item) {
         log.trace(this.getClass().getSimpleName() + ".put = " + item);
         try {
-            Thread.sleep(5000);
+            Thread.sleep(3000);
         } catch (InterruptedException e) {
         }
         called();
-        return 1;
+        return 2;
     }
 
     @Override
     public String get(int id) {
         called();
         log.trace(this.getClass().getSimpleName() + ".get = " + id);
-        return "null 1";
+        return "null 2";
     }
 
     @Override
