+++ src/test/java/com/github/lukaszbudnik/guice/multibindinginterceptor/ioc/test/TestModule.java
@@ -1,4 +1,4 @@
-package io.github.lukaszbudnik.guice.multibindinginterceptor.ioc.test;
+package com.github.lukaszbudnik.guice.multibindinginterceptor.ioc.test;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.Multibinder;
